+++ src/services/Elastic.Changelog/Configuration/ChangelogConfigurationLoader.cs
@@ -521,6 +521,7 @@ public class ChangelogConfigurationLoader(ILoggerFactory logFactory, IConfigurat
 						Repo = kvp.Value.Repo,
 						Owner = kvp.Value.Owner,
 						HideFeatures = kvp.Value.HideFeatures?.Values,
+						ShowReleaseDates = kvp.Value.ShowReleaseDates,
 						Source = kvp.Value.Source
 					});
 		}
@@ -534,6 +535,7 @@ public class ChangelogConfigurationLoader(ILoggerFactory logFactory, IConfigurat
 			Repo = yaml.Repo,
 			Owner = yaml.Owner,
 			LinkAllowRepos = linkAllowRepos,
+			ShowReleaseDates = yaml.ShowReleaseDates ?? false,
 			Profiles = profiles
 		};
 	}